requirements.txt
Flask==2.1.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-MySQLdb==1.0.1
Werkzeug==2.1.1
watchdog==2.1.9
Pyrebase4==4.5.0
functions-framework==3.1.0
requests==2.27.1
requests-oauthlib==1.3.1
sendgrid==6.9.7
urllib3==1.26.9
Werkzeug==2.1.1

